# openapi/openapi.yaml
openapi: 3.0.3

info:
  title: Marketplace API
  version: 1.0.0
  description: API contract for the Marketplace course project.

security: []

servers:
  - url: http://localhost:3000

paths:
  /products:
    get:
      operationId: getProducts
      summary: Get products list
      description: Returns a paginated list of marketplace products.
      parameters:
        - in: query
          name: limit
          required: false
          description: Maximum number of products to return.
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10

        - in: query
          name: cursor
          required: false
          description: Opaque pagination token returned by the previous request. Clients must not parse or modify it.
          schema:
            type: string

      responses:
        '200':
          description: Products list returned successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductListResponse'

        '400':
          $ref: '#/components/responses/BadRequest'

  /products/{id}:
    get:
      operationId: getProductById
      summary: Get product by id
      description: Returns a single marketplace product.
      parameters:
        - in: path
          name: id
          required: true
          description: Product identifier.
          schema:
            type: integer
            minimum: 1

      responses:
        '200':
          description: Product returned successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'

        '404':
          $ref: '#/components/responses/NotFound'

  /orders:
    get:
      operationId: getOrders
      summary: Get orders list
      description: Returns a paginated list of marketplace orders.
      parameters:
        - in: query
          name: limit
          required: false
          description: Maximum number of orders to return.
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10

        - in: query
          name: cursor
          required: false
          description: Opaque pagination token returned by the previous request. Clients must not parse or modify it.
          schema:
            type: string

      responses:
        '200':
          description: Orders list returned successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderListResponse'

        '400':
          $ref: '#/components/responses/BadRequest'

    post:
      operationId: createOrder
      summary: Create order
      description: Creates a new marketplace order.
      parameters:
        - in: header
          name: Idempotency-Key
          required: true
          description: Repeating a request with the same Idempotency-Key and the same request body must return the same response as the original request.
          schema:
            type: string
            minLength: 1

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'

      responses:
        '201':
          description: Order created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'

        '400':
          $ref: '#/components/responses/BadRequest'

  /orders/{id}:
    get:
      operationId: getOrderById
      summary: Get order by id
      description: Returns a single marketplace order.
      parameters:
        - in: path
          name: id
          required: true
          description: Order identifier.
          schema:
            type: integer
            minimum: 1

      responses:
        '200':
          description: Order returned successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'

        '404':
          $ref: '#/components/responses/NotFound'

components:
  schemas:
    Product:
      type: object
      required:
        - id
        - name
        - price_cents
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: Mechanical Keyboard
        price_cents:
          type: integer
          example: 260000

    OrderItem:
      type: object
      required:
        - product_id
        - quantity
      properties:
        product_id:
          type: integer
          minimum: 1
          example: 1
        quantity:
          type: integer
          minimum: 1
          example: 2

    CreateOrderRequest:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/OrderItem'

    Order:
      type: object
      required:
        - id
        - items
        - total_cents
        - status
      properties:
        id:
          type: integer
          example: 1
        items:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/OrderItem'
        total_cents:
          type: integer
          example: 520000
        status:
          type: string
          example: created

    ProductListResponse:
      type: object
      required:
        - items
        - next_cursor
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Product'
        next_cursor:
          type: string
          nullable: true
          description: Opaque cursor for the next page. Null means there are no more pages.
          example: null

    OrderListResponse:
      type: object
      required:
        - items
        - next_cursor
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Order'
        next_cursor:
          type: string
          nullable: true
          description: Opaque cursor for the next page. Null means there are no more pages.
          example: null

    Problem:
      type: object
      required:
        - type
        - title
        - status
        - detail
        - instance
      properties:
        type:
          type: string
          format: uri
          description: URI identifying the problem type.
          example: https://example.com/problems/validation-error

        title:
          type: string
          description: Short human-readable problem title.
          example: Validation error

        status:
          type: integer
          description: HTTP status code.
          example: 400

        detail:
          type: string
          description: Human-readable explanation of what went wrong.
          example: Request validation failed.

        instance:
          type: string
          description: URI identifying this particular occurrence of the problem.
          example: /orders

  responses:
    BadRequest:
      description: Invalid request.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'

    NotFound:
      description: Resource not found.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'